requirements.txt
streamlit==1.28.0
langchain
langchain-google-genai
langchain-core
langchain-text-splitters
chromadb
python-dotenv
numpy
pydantic
chunking_evaluation @ git+https://github.com/brandonstarxel/chunking_evaluation.git@d451fc4cf56e417b755994b4ca5212fd5057c0d2
